requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
contextlib3==3.10.0
discord==1.7.3
discord.py==1.7.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
MemeLib==0.1.5
multidict==5.1.0
pipfreeze==2.0.3
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
python-decouple==3.4
python-dotenv==0.18.0
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Wikipedia-API==0.5.4
yarl==1.6.3
youtube-dl==2021.6.6
youtube-search==2.1.0
